pyproject.toml
[tool.poetry]
name = "ecom-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^4.1.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}
slugify = "^0.0.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
alembic = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
